🚚 Move integer arbitraries into /arbitrary
