+++ src/check/arbitrary/IpArbitrary.ts
@@ -1,7 +1,7 @@
 import { array } from './ArrayArbitrary';
 import { constantFrom } from './ConstantArbitrary';
 import { Arbitrary } from './definition/Arbitrary';
-import { nat } from './IntegerArbitrary';
+import { nat } from '../../arbitrary/nat';
 import { oneof } from './OneOfArbitrary';
 import { hexaString } from './StringArbitrary';
 import { tuple } from './TupleArbitrary';
