+++ test/unit/check/arbitrary/DateArbitrary.utest.spec.ts
@@ -4,13 +4,11 @@ import { mocked } from 'ts-jest/utils';
 import * as fc from '../../../../lib/fast-check';
 import { ArbitraryWithShrink } from '../../../../src/check/arbitrary/definition/ArbitraryWithShrink';
 
-jest.mock('../../../../src/check/arbitrary/IntegerArbitrary');
-import * as _IntegerArbitraryMock from '../../../../src/check/arbitrary/IntegerArbitrary';
+jest.mock('../../../../src/arbitrary/integer');
+import * as _IntegerMock from '../../../../src/arbitrary/integer';
 import { arbitraryFor } from './generic/ArbitraryBuilder';
 
-const IntegerArbitraryMock: {
-  integer: (min: number, max: number) => ArbitraryWithShrink<number>;
-} = _IntegerArbitraryMock;
+const IntegerArbitraryMock: { integer: (min: number, max: number) => ArbitraryWithShrink<number> } = _IntegerMock;
 
 const mrng = () => stubRng.mutable.nocall();
