+++ test/unit/check/arbitrary/FunctionArbitrary.spec.ts
@@ -2,7 +2,7 @@ import * as fc from '../../../../lib/fast-check';
 
 import { func, compareFunc, compareBooleanFunc } from '../../../../src/check/arbitrary/FunctionArbitrary';
 import { context } from '../../../../src/check/arbitrary/ContextArbitrary';
-import { integer } from '../../../../src/check/arbitrary/IntegerArbitrary';
+import { integer } from '../../../../src/arbitrary/integer';
 import { hasCloneMethod, cloneMethod } from '../../../../src/check/symbols';
 import { hash } from '../../../../src/utils/hash';
 import { stringify } from '../../../../src/utils/stringify';
