+++ test/unit/check/arbitrary/OneOfArbitrary.itest.spec.ts
@@ -3,7 +3,7 @@ import * as fc from '../../../../lib/fast-check';
 import { Arbitrary } from '../../../../src/check/arbitrary/definition/Arbitrary';
 import { Shrinkable } from '../../../../src/check/arbitrary/definition/Shrinkable';
 import { constant } from '../../../../src/check/arbitrary/ConstantArbitrary';
-import { integer } from '../../../../src/check/arbitrary/IntegerArbitrary';
+import { integer } from '../../../../src/arbitrary/integer';
 import { oneof } from '../../../../src/check/arbitrary/OneOfArbitrary';
 import { Random } from '../../../../src/random/generator/Random';
 import { stream } from '../../../../src/stream/Stream';
