+++ test/unit/check/arbitrary/UuidArbitrary.utest.spec.ts
@@ -3,16 +3,16 @@ import * as stubRng from '../../stubs/generators';
 import { mocked } from 'ts-jest/utils';
 import { ArbitraryWithShrink } from '../../../../src/check/arbitrary/definition/ArbitraryWithShrink';
 
-jest.mock('../../../../src/check/arbitrary/IntegerArbitrary');
+jest.mock('../../../../src/arbitrary/integer');
+jest.mock('../../../../src/arbitrary/nat');
 jest.mock('../../../../src/check/arbitrary/TupleArbitrary');
-import * as _IntegerArbitraryMock from '../../../../src/check/arbitrary/IntegerArbitrary';
+import * as _IntegerMock from '../../../../src/arbitrary/integer';
+import * as _NatMock from '../../../../src/arbitrary/nat';
 import * as TupleArbitraryMock from '../../../../src/check/arbitrary/TupleArbitrary';
 import { arbitraryFor } from './generic/ArbitraryBuilder';
 
-const IntegerArbitraryMock: {
-  integer: (min: number, max: number) => ArbitraryWithShrink<number>;
-  nat: (max: number) => ArbitraryWithShrink<number>;
-} = _IntegerArbitraryMock;
+const IntegerMock: { integer: (min: number, max: number) => ArbitraryWithShrink<number> } = _IntegerMock;
+const NatMock: { nat: (max: number) => ArbitraryWithShrink<number> } = _NatMock;
 
 const mrng = () => stubRng.mutable.nocall();
 
@@ -23,7 +23,8 @@ describe('UuidArbitrary', () => {
     });
     it('Should be able to build the smallest uuid', () => {
       // Arrange
-      const { nat, integer } = mocked(IntegerArbitraryMock);
+      const { integer } = mocked(IntegerMock);
+      const { nat } = mocked(NatMock);
       const { tuple } = mocked(TupleArbitraryMock);
       nat.mockImplementation(() => arbitraryFor([{ value: 0 }, { value: 0 }, { value: 0 }]));
       integer.mockImplementation((a, _b) => arbitraryFor([{ value: a }]));
@@ -41,7 +42,8 @@ describe('UuidArbitrary', () => {
     });
     it('Should be able to build the largest uuid', () => {
       // Arrange
-      const { nat, integer } = mocked(IntegerArbitraryMock);
+      const { integer } = mocked(IntegerMock);
+      const { nat } = mocked(NatMock);
       const { tuple } = mocked(TupleArbitraryMock);
       nat.mockImplementation((a) => arbitraryFor([{ value: a }, { value: a }, { value: a }]));
       integer.mockImplementation((a, b) => arbitraryFor([{ value: b }]));
@@ -64,7 +66,8 @@ describe('UuidArbitrary', () => {
     });
     it('Should be able to build the smallest versioned uuid', () => {
       // Arrange
-      const { nat, integer } = mocked(IntegerArbitraryMock);
+      const { integer } = mocked(IntegerMock);
+      const { nat } = mocked(NatMock);
       const { tuple } = mocked(TupleArbitraryMock);
       nat.mockImplementation(() => arbitraryFor([{ value: 0 }, { value: 0 }]));
       integer.mockImplementation((a, _b) => arbitraryFor([{ value: a }]));
@@ -82,7 +85,8 @@ describe('UuidArbitrary', () => {
     });
     it('Should be able to build the largest versioned uuid', () => {
       // Arrange
-      const { nat, integer } = mocked(IntegerArbitraryMock);
+      const { integer } = mocked(IntegerMock);
+      const { nat } = mocked(NatMock);
       const { tuple } = mocked(TupleArbitraryMock);
       nat.mockImplementation((a) => arbitraryFor([{ value: a }, { value: a }]));
       integer.mockImplementation((a, b) => arbitraryFor([{ value: b }]));
